Some changes in migration

+++ database/migrations/2023_05_17_102450_create_donations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id()->autoIncrement();
             $table->string('donatorname');
             $table->string('email')->unique();
-            $table->integer('amount')->unsigned();
+            $table->decimal('amount', 8, 2)->unsigned();
             $table->text('message')->nullable();
             $table->date('date');
             $table->timestamps();
